# public/openapi.yaml
openapi: 3.1.0
info:
  title: DataMaster SSO Authentication API
  version: 1.0.0
  description: |
    # DataMaster SSO Flow - OAuth 2.0 + PKCE

    **日本語**: このAPIは、DataMasterアプリケーションとSmart iMATE IDプロバイダー間のOAuth 2.0 + PKCE認証フローを実装します。AWS Cognitoをバックエンドとして使用し、マルチテナント対応のセキュアな認証を提供します。

    **Tiếng Việt**: API này triển khai luồng xác thực OAuth 2.0 + PKCE giữa ứng dụng DataMaster và nhà cung cấp định danh Smart iMATE. Sử dụng AWS Cognito làm backend và cung cấp xác thực an toàn hỗ trợ đa tenant.

    ## Security Features / セキュリティ機能 / Tính năng bảo mật

    - **PKCE (RFC 7636)**: Prevents authorization code interception attacks
      - **日本語**: 認可コード傍受攻撃を防止
      - **Tiếng Việt**: Ngăn chặn tấn công đánh cắp mã ủy quyền

    - **State Parameter**: CSRF protection via signed JWT
      - **日本語**: 署名付きJWTによるCSRF保護
      - **Tiếng Việt**: Bảo vệ CSRF qua JWT đã ký

    - **Multi-Tenant Isolation**: Tenant-scoped cookies and validation
      - **日本語**: テナント単位のCookieと検証
      - **Tiếng Việt**: Cookie và xác thực theo phạm vi tenant

    ## Architecture / アーキテクチャ / Kiến trúc

    ```
    DataMaster (Client)
           ↕ OAuth 2.0 + PKCE
    Smart iMATE (Identity Provider)
           ↕ USER_PASSWORD_AUTH
    AWS Cognito (User Pool)
    ```
servers:
  - url: http://localhost:3000
    description: |
      **日本語**: DataMaster開発環境
      **Tiếng Việt**: Môi trường phát triển DataMaster
  - url: http://localhost:8080
    description: |
      **日本語**: Smart iMATE IDプロバイダー
...
paths:
...
  /api/auth/signin:
    get:
      tags: [Phase 1]
      summary: Initiate OAuth Authorization
      description: |
        **日本語説明**:
        OAuth 2.0 + PKCE認証フローを開始します。DataMasterがPKCEパラメータ（code_verifier、code_challenge）と署名付きstateトークンを生成し、Smart iMATEの認可エンドポイントにリダイレクトします。

        **セキュリティメカニズム**:
        - **code_verifier**: 32バイトのランダム値をbase64urlエンコード（256ビット）
        - **code_challenge**: code_verifierのSHA-256ハッシュをbase64urlエンコード
        - **state**: テナント、code_verifier、callbackUrlを含む署名付きJWT（HS256）

        **Cookieに保存**:
        - `oauth_state_{tenant}`: stateトークン（10分有効）
        - `oauth_code_verifier_{tenant}`: PKCEのcode_verifier
        - `oauth_tenant_{tenant}`: テナント識別子

        **Giải thích tiếng Việt**:
        Bắt đầu luồng xác thực OAuth 2.0 + PKCE. DataMaster tạo các tham số PKCE (code_verifier, code_challenge) và token state đã ký, sau đó chuyển hướng đến endpoint ủy quyền Smart iMATE.

        **Cơ chế bảo mật**:
        - **code_verifier**: Giá trị ngẫu nhiên 32 byte mã hóa base64url (256-bit)
        - **code_challenge**: Hash SHA-256 của code_verifier mã hóa base64url
        - **state**: JWT đã ký (HS256) chứa tenant, code_verifier, callbackUrl

        **Lưu trong Cookie**:
        - `oauth_state_{tenant}`: Token state (hiệu lực 10 phút)
        - `oauth_code_verifier_{tenant}`: code_verifier PKCE
        - `oauth_tenant_{tenant}`: Định danh tenant

      parameters:
        - name: tenant
          in: query
          required: true
          description: |
            **日本語**: テナント識別子（例: "kubota"）
            **Tiếng Việt**: Định danh tenant (ví dụ: "kubota")
          schema:
            type: string
            example: kubota

      responses:
        '302':
          description: Redirect to Smart iMATE authorization endpoint
          headers:
            Location:
              description: Smart iMATE OAuth authorize URL
              schema:
                type: string
                example: http://localhost:8080/oauth2/authorize?client_id=1a2b3c4d&redirect_uri=http://localhost:3000/api/auth/callback&response_type=code&scope=openid+email+profile&state=eyJhbGc...&code_challenge=E9Melhoa2OwvFrEMTJguCHaoeK1t8URWbuGJSstw-cM&code_challenge_method=S256
            Set-Cookie:
              description: OAuth temporary cookies
              schema:
                type: array
                items:
                  type: string
                example:
                  - oauth_state_kubota=eyJhbGc...; Max-Age=600; HttpOnly; Secure; SameSite=Lax
                  - oauth_code_verifier_kubota=dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk; Max-Age=600; HttpOnly; Secure; SameSite=Lax
                  - oauth_tenant_kubota=kubota; Max-Age=600; HttpOnly; Secure; SameSite=Lax
...
  /dashboard:
    get:
      tags: [Phase 5-7]
      summary: Authenticated Dashboard
      description: |
        **日本語説明**:
        認証済みユーザーのダッシュボード。セッションJWT Cookieで保護されています。

        **表示内容**:
        - ユーザー情報（名前、メール、サブ）
        - テナントコンテキスト
        - Smart iMATEへのSSOナビゲーションボタン

        **セッション検証**:
        - Cookie `session` JWT検証（HS256）
        - 有効期限チェック（30日）
        - テナント境界確認

        **Giải thích tiếng Việt**:
        Dashboard người dùng đã xác thực. Được bảo vệ bởi Session JWT Cookie.

        **Nội dung hiển thị**:
        - Thông tin người dùng (tên, email, sub)
        - Ngữ cảnh tenant
        - Nút điều hướng SSO đến Smart iMATE

        **Xác thực session**:
        - Xác thực Cookie `session` JWT (HS256)
        - Kiểm tra hết hạn (30 ngày)
        - Xác nhận ranh giới tenant

      security:
        - sessionCookie: []

      responses:
        '200':
          description: |
            **日本語**: ダッシュボードページ表示
            **Tiếng Việt**: Hiển thị trang dashboard
          content:
            text/html:
              schema:
                type: string

        '401':
          description: |
            **日本語**: 未認証、ログインページへリダイレクト
            **Tiếng Việt**: Chưa xác thực, chuyển hướng đến trang đăng nhập
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        **日本語**: Cognito Access Token JWT（1時間有効）
        **Tiếng Việt**: JWT Cognito Access Token (hiệu lực 1 giờ)
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        **日本語**: セッションJWT Cookie（HS256署名、30日有効）
        **Tiếng Việt**: Session JWT Cookie (chữ ký HS256, hiệu lực 30 ngày)